Grand Unification
After looking at all the redundant code in the four individual tests that make up the shopping test, this on test combines all of them into a single flow. In addition, it tests whether the form is validating the empty fields properly.

Accordingly, the smoke test plan has been updated. And for this script in the smoke test, we've simplified the language since we now have a familiarity with the Capybara DSL and can answer some of the questions that were posed in the original smoke test plan.