First back-office test: "Order Processing"
This test follows the test plan by simulating a purchase, then logging into the application, reviewing the order; editing its payment type; then destroying it.

This commit has a lot of changes that reflect a sort of "upgrade" to our testing process.

Firstly, this test ("backoffice processes an order") is using the selenium driver by setting "js: true" in the test block.

Since we're now using the selenium driver, we need to also use the database_cleaner gem to help us with non-transactional fixtures/ORM. This is because a single test spans several web/db transactions. (Gemfile, Gemfile.lock, spec_helper.rb, database_cleaner.rb)

SQLite3 is a pain to use with selenium (or at all really), so we're changing the test environment to use the mysql2 adapter, instead. (database.yml)

The selenium driver is a little more aggressive about failures (because it's watching an actual browser -- Firefox), so we found out that our products.yml fixture is buggy (specifically, when we display the catalog, several records were 404'ing the image_url, causing the test to fail. This might be avoidable, but that's as an exercise to the reader -- just revert to the prior version of the products.yml file and you're off to the races.