Back-office test for managing Products
The RSpec script follows the test plan by the letter, including a sneaky (hacky?) counting of rows in the database instead of trying to do any sort of clever xpath/css table row counting.

This also includes an image that our test uses, seeing as how the test generates a routing error without it (because it's a web-driven test). Again, there may be a way to turn that off, but that's up to you. Just delete the image and run the test to see what happens.